chore(gfx): fixed GCC build errors

+++ src/lepus/utility/types/Quaternion.h
@@ -43,7 +43,7 @@ namespace lepus
 	    /// @param angle Rotation angle around the axis
 	    Quaternion(float axisX, float axisY, float axisZ, float angle)
 	    {
-		float const q[] = {axisX * sinf(angle / 2.f), axisY * sinf(angle / 2.f), axisZ * sinf(angle / 2.f), cos(angle / 2.f)};
+		float const q[] = {axisX * sinf(angle / 2.f), axisY * sinf(angle / 2.f), axisZ * sinf(angle / 2.f), cosf(angle / 2.f)};
 		init((float*)q);
 	    }
